Refactor routers
