+++ .prettierrc
@@ -3,6 +3,7 @@
   "trailingComma": "none",
   "printWidth": 80,
   "semi": false,
+  "useTabs": false,
   "tabWidth": 2,
   "overrides": [
     {
