+++ views/partials/tabs.hbs
@@ -1,10 +1,7 @@
 <div class="container">
     <div class="tabs is-boxed">
         <ul>
-            <li class="{{#eq title "Dashboard"}}is-active{{/eq}}"><a href="/admin/"><i
-                        class="fas fa-tachometer-alt"></i>&ensp;Dashboard</a></li>
-            <li class="{{#eq title "Auto Email"}}is-active{{/eq}}"><a href="/admin/email"><i
-                        class="fas fa-mail-bulk"></i>&ensp;Auto Email</a></li>
+            <li class="{{#eq title "Dashboard"}}is-active{{/eq}}"><a href="/admin/"><i class="fas fa-tachometer-alt"></i>&ensp;Dashboard</a></li>
         </ul>
     </div>
 </div>
